CAPACITY -> MAXIMUM_SIZE
std::vector has both capacity() and max_size(). The former typically
changes as the vector grows. For FixedVector they happen to be the same,
but MAX_SIZE better reflects the purpose of the template parameter.
For example, when checking if there is space for more elements, the check
should be against "max size" and not "capacity".
Furthermore, std::map/std::set have max_size() but not capacity().

Defensively use MAXIMUM_SIZE instead of MAX_SIZE to reduce chance of
collision with macros.