[SkeletalRichEnum] Fix ordinal() and value_of(underlying_int)
Ordinal was erroneously returning the enum_integer as opposed
to the enum_index, which only works for [0,N) enums.
Fix it through magic_enum, also SkeletalRichEnumLite no longer
provides `ordinal() due to this.

value_of(int) was also erroneously assuming that the enum
is [0, N). Fix it and change variable names/types to better
express that this is the enum_integer and not the enum_index