Add FixedRobinhoodHashtable based on `ankerl::UnorderedDense`
This is an implmentation of a hashtable based on Robin Hood hashing with
backwards shift deletion. The code was initially based on Martin
Ankerl's `UnorderedDense` and still uses an identical algorithm, but
nearly all of it has been rewritten. This is intentionally a bare-bones
API, with only the functions required to implement a full-fat Map API.