Rewrite nontrivial copy and move for FixedDoublyLinkedList
This was the source of a bug in FixedUnorderedMap for
non-TriviallyCopyably types caused by retaining indices into the
LinkedList's underlying storage, but the old copy/move logic would
change the locations of the stored values. The new operations instead
place the values at the same indices of the destination storage.